chore: add terminal page and wouter
