+++ examples/package.json
@@ -36,6 +36,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-error-boundary": "^6.0.0",
+    "wouter": "^3.7.1",
     "zustand": "^5.0.3"
   }
 }
